This commit was generated by cvs2svn to compensate for changes in r15289, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@15290 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/video/render2x2.c
@@ -56,7 +56,8 @@ void render_08_2x2_04(const DWORD *colortab, const BYTE *src, BYTE *trg,
         wfast = 0;
         wend = 0;
     } else {
-        wstart=8-((unsigned int)trg & 7); /* alignment: 8 pixels*/
+        /* alignment: 8 pixels*/
+        wstart = (unsigned int)(8 - ((unsigned long)trg & 7));
         wfast = (width - wstart) >> 3; /* fast loop for 8 pixel segments*/
         wend = (width - wstart) & 0x07; /* do not forget the rest*/
     }
@@ -146,7 +147,8 @@ void render_16_2x2_04(const DWORD *colortab, const BYTE *src, BYTE *trg,
         wfast = 0;
         wend = 0;
     } else {
-        wstart = 8 - ((unsigned int)trg & 7); /* alignment: 8 pixels*/
+        /* alignment: 8 pixels*/
+        wstart = (unsigned int)(8 - ((unsigned long)trg & 7));
         wfast = (width - wstart) >> 3; /* fast loop for 8 pixel segments*/
         wend = (width - wstart) & 0x07; /* do not forget the rest*/
     }
@@ -237,7 +239,8 @@ void render_24_2x2_04(const DWORD *colortab, const BYTE *src, BYTE *trg,
         wfast = 0;
         wend = 0;
     } else {
-        wstart = 4 - ((unsigned int)trg & 3); /* alignment: 4 pixels*/
+        /* alignment: 4 pixels*/
+        wstart = (unsigned int)(4 - ((unsigned long)trg & 3));
         wfast = (width - wstart) >> 2; /* fast loop for 4 pixel segments*/
         wend = (width - wstart) & 0x03; /* do not forget the rest*/
     }
@@ -430,7 +433,8 @@ void render_32_2x2_04(const DWORD *colortab, const BYTE *src, BYTE *trg,
         wfast = 0;
         wend = 0;
     } else {
-        wstart = 8 - ((unsigned int)trg & 7); /* alignment: 8 pixels*/
+        /* alignment: 8 pixels*/
+        wstart = (unsigned int)(8 - ((unsigned long)trg & 7));
         wfast = (width - wstart) >> 3; /* fast loop for 8 pixel segments*/
         wend = (width - wstart) & 0x07; /* do not forget the rest*/
     }
@@ -555,7 +559,8 @@ void render_08_2x2_08(const DWORD *colortab, const BYTE *src, BYTE *trg,
         wfast = 0;
         wend = 0;
     } else {
-        wstart = 8 - ((unsigned int)trg & 7); /* alignment: 8 pixels*/
+        /* alignment: 8 pixels*/
+        wstart = (unsigned int)(8 - ((unsigned long)trg & 7));
         wfast = (width - wstart) >> 3; /* fast loop for 8 pixel segments*/
         wend = (width - wstart) & 0x07; /* do not forget the rest*/
     }
@@ -566,7 +571,8 @@ void render_08_2x2_08(const DWORD *colortab, const BYTE *src, BYTE *trg,
         tmptrg = (WORD *)trg;
         if ((y & 1) || doublescan) {
             if (wfirst) {
-                *((BYTE *)tmptrg) = (BYTE)colortab[*tmpsrc++ | (*tmppre++ << 4)];
+                *((BYTE *)tmptrg) = (BYTE)colortab[*tmpsrc++
+                                    | (*tmppre++ << 4)];
                 tmptrg = (WORD *)(((BYTE *)tmptrg) + 1);
             }
             for (x = 0; x < wstart; x++) {
@@ -652,7 +658,8 @@ void render_16_2x2_08(const DWORD *colortab, const BYTE *src, BYTE *trg,
         wfast = 0;
         wend = 0;
     } else {
-        wstart = 8 - ((unsigned int)trg & 7); /* alignment: 8 pixels*/
+        /* alignment: 8 pixels*/
+        wstart = (unsigned int)(8 - ((unsigned long)trg & 7));
         wfast = (width - wstart) >> 3; /* fast loop for 8 pixel segments*/
         wend = (width - wstart) & 0x07; /* do not forget the rest*/
     }
@@ -663,7 +670,8 @@ void render_16_2x2_08(const DWORD *colortab, const BYTE *src, BYTE *trg,
         tmptrg = (DWORD *)trg;
         if ((y & 1) || doublescan) {
             if (wfirst) {
-                *((WORD *)tmptrg) = (WORD)colortab[*tmpsrc++ | (*tmppre++ << 4)];
+                *((WORD *)tmptrg) = (WORD)colortab[*tmpsrc++
+                                    | (*tmppre++ << 4)];
                 tmptrg = (DWORD *)(((WORD *)tmptrg) + 1);
             }
             for (x = 0; x < wstart; x++) {
@@ -750,7 +758,8 @@ void render_24_2x2_08(const DWORD *colortab, const BYTE *src, BYTE *trg,
         wfast = 0;
         wend = 0;
     } else {
-        wstart = 4 - ((unsigned int)trg & 3); /* alignment: 4 pixels*/
+        /* alignment: 4 pixels*/
+        wstart = (unsigned int)(4 - ((unsigned long)trg & 3));
         wfast = (width - wstart) >> 2; /* fast loop for 4 pixel segments*/
         wend = (width - wstart) & 0x03; /* do not forget the rest*/
     }
@@ -950,7 +959,8 @@ void render_32_2x2_08(const DWORD *colortab, const BYTE *src, BYTE *trg,
         wfast = 0;
         wend = 0;
    } else {
-        wstart = 4 - ((unsigned int)trg & 3); /* alignment: 4 pixels*/
+        /* alignment: 4 pixels*/
+        wstart = (unsigned int)(4 - ((unsigned long)trg & 3));
         wfast = (width - wstart) >> 2; /* fast loop for 4 pixel segments*/
         wend = (width - wstart) & 0x03; /* do not forget the rest*/
     }
