+++ vice/src/video/render2x2pal.c
@@ -66,8 +66,8 @@ void render_16_2x2_pal(const DWORD *colortab, const BYTE *src, BYTE *trg,
     SDWORD l, u, v;
     DWORD red, grn, blu, red2, grn2, blu2;
 
-    src=src + pitchs * ys + xs - 2;
-    trg=trg + pitcht * yt + (xt << 1);
+    src = src + pitchs * ys + xs - 2;
+    trg = trg + pitcht * yt + (xt << 1);
     yys = (ys << 1) | (yt & 1);
     wfirst = xt & 1;
     width -= wfirst;
@@ -78,7 +78,8 @@ void render_16_2x2_pal(const DWORD *colortab, const BYTE *src, BYTE *trg,
         wfast = 0;
         wend = 0;
     } else {
-        wstart = 8 - ((unsigned int)trg & 7); /* alignment: 8 pixels*/
+        /* alignment: 8 pixels*/
+        wstart = (unsigned int)(8 - ((unsigned long)trg & 7));
         wfast = (width - wstart) >> 3; /* fast loop for 8 pixel segments*/
         wend = (width - wstart) & 0x07; /* do not forget the rest*/
     }
@@ -322,8 +323,8 @@ void render_32_2x2_pal(const DWORD *colortab, const BYTE *src, BYTE *trg,
     SDWORD l, u, v;
     DWORD red, grn, blu, red2, grn2, blu2;
 
-    src=src + pitchs * ys + xs - 2;
-    trg=trg + pitcht * yt + (xt << 2);
+    src = src + pitchs * ys + xs - 2;
+    trg = trg + pitcht * yt + (xt << 2);
     yys = (ys << 1) | (yt & 1);
     wfirst = xt & 1;
     width -= wfirst;
@@ -334,7 +335,8 @@ void render_32_2x2_pal(const DWORD *colortab, const BYTE *src, BYTE *trg,
         wfast = 0;
         wend = 0;
     } else {
-        wstart = 8 - ((unsigned int)trg & 7); /* alignment: 8 pixels*/
+        /* alignment: 8 pixels*/
+        wstart = (unsigned int)(8 - ((unsigned long)trg & 7));
         wfast = (width - wstart) >> 3; /* fast loop for 8 pixel segments*/
         wend = (width - wstart) & 0x07; /* do not forget the rest*/
     }
